requirements.txt
# python_requires = ">=3.11,<3.12"
# ETL
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.11.0

# Base de datos
supabase==2.15.1
httpx>=0.27.0
python-dotenv==1.0.0

# Procesamiento de datos
pandas==2.2.0
numpy==1.26.4
statsmodels>=0.14.0   # requerido por plotly trendline="ols"

# Dashboard / Modelo
scikit-learn==1.4.0
streamlit==1.31.0
plotly==5.18.0
shap==0.44.0
python-dateutil>=2.8.2
